pyproject.toml
[tool.poetry]
name = "stocks-api"
version = "0.1.0"
description = ""
authors = ["Roman Sydoruk <sydoruk89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.2"
djangorestframework = "^3.12.1"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
